Yet more sdl hacking, region memory leak fix, mlt_position changed to int32_t, experimental hash in properties
git-svn-id: https://mlt.svn.sourceforge.net/svnroot/mlt/trunk/mlt@184 d19143bc-622f-0410-bfdd-b5b2a6649095

+++ src/framework/mlt_types.h
@@ -33,7 +33,7 @@ typedef enum
 }
 mlt_whence;
 
-typedef int64_t mlt_position;
+typedef int32_t mlt_position;
 typedef struct mlt_frame_s *mlt_frame, **mlt_frame_ptr;
 typedef struct mlt_properties_s *mlt_properties;
 typedef struct mlt_service_s *mlt_service;
